requirements.txt
absl-py==0.5.0
astor==0.7.1
awscli==1.16.5
awsebcli==3.14.4
backports.functools-lru-cache==1.5
backports.ssl-match-hostname==3.5.0.1
backports.weakref==1.0.post1
blessed==1.15.0
botocore==1.11.5
cached-property==1.4.3
cement==2.8.2
certifi==2018.8.13
chardet==3.0.4
click==6.7
colorama==0.3.9
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cycler==0.10.0
Django==1.11.15
django-bootstrap-form==3.4
django-bootstrap3==11.0.0
django-dashing==0.3.3
django-db-logger==0.1.6
django-filter==1.1.0
django-graphos==0.3.41
django-rest-auth==0.9.3
django-sb-admin==0.3.7
django-tables2==2.0.1
djangorestframework==3.8.2
docker==3.5.0
docker-compose==1.21.2
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
idna==2.6
ipaddress==1.0.22
itypes==1.1.0
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mock==2.0.0
mysqlclient==1.3.13
numpy==1.15.0
pathspec==0.5.5
pbr==4.2.0
Pillow==5.2.0
pkg-resources==0.0.0
protobuf==3.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pyasn1==0.4.4
pyasn1-modules==0.2.2
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
semantic-version==2.5.0
six==1.11.0
sorl-thumbnail==12.4.1
subprocess32==3.5.2
tabulate==0.7.5
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
texttable==0.9.1
unicorn==1.0.1
uritemplate==3.0.0
urllib3==1.23
wcwidth==0.1.7
websocket-client==0.51.0
Werkzeug==0.14.1
whitenoise==4.1
